+++ old_code/analyze_database/check_db_sequence_number.py
