+++ old_code/analyze_database/check_db_subspc.py
