+++ old_code/analyze_database/find_duplicate_seq.py
