+++ old_code/analyze_database/find_partial_seq.py
