+++ old_code/analyze_database/get_mugil_cephalus_seq.py
