+++ old_code/analyze_database/get_seq_descip.py
