+++ old_code/deprecated_main_code/(deprecated)analysis_manager.py
