+++ old_code/deprecated_main_code/(deprecated)read_all.py
